% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{autoencode}
\alias{autoencode}
\title{Autoencoder
Implementation of autoencoder via \code{mxnet} library. It can be used for denoising and dimension reduction.}
\usage{
autoencode(X.train, X.test = NULL, nl = 3, N.hidden = c(100),
  actType = c("sigmoid", "tanh", "relu"), lambda, beta, rho, epsilon,
  optim.method = c("BFGS", "L-BFGS-B", "CG"),
  rel.tol = sqrt(.Machine$double.eps), max.iterations = 2000,
  rescale.flag = c(F, T), rescaling.offset = 0.001)
}
\description{
Autoencoder
Implementation of autoencoder via \code{mxnet} library. It can be used for denoising and dimension reduction.
}
