% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{visulizeDigits}
\alias{visulizeDigits}
\title{Visualize the digits. It turns a vector with 784 dimensions into a 28*28 pickture. It needs 'softmaxreg' package to be installed.}
\usage{
visulizeDigits(x)
}
\arguments{
\item{x}{A vector with 784 dimensions.}
}
\description{
Visualize the digits. It turns a vector with 784 dimensions into a 28*28 pickture. It needs 'softmaxreg' package to be installed.
}
